+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+Added proper configuration for default scoreboard  
+Added `^` (exponential) operator  
+Re-implement extension's configuration  
+Fixed "Insert math expression split equation where it's not suppose to"  
+Fixed "Planed Features typo"
+
 ## 1.0.3
 Fixed "error throw when first install"